Save different instances of subcircuits separately while keeping the format backward compatible
Although different instances of subcircuits starts identical,
they corresponds to distinct runtime object and their internal
may deviate from each other due to edits done on each subcircuits.

The current format disgard all the changes except for the first subcircuit
and apply the edit on the first subcircuit to all instances during loading.

This change modifies the saving format while keeping the file backward compatible
so that the new file can still be openned by old code, with only minor display differences.
This is done by automatically generate new celltype for each instances of the subcircuit
on saving and saving the old value that is only useful for display in a separate attribute
`disp_celltype`. This automatic name generation is also only done when there are
multiple instances of the same cell type so that we can keep generating names shorter
and the file as closed to the old format as possible.